feat(merge shell with daemon): mergin shell with daemon
